app/requirements.txt
pandas
numpy
matplotlib
seaborn
langchain==0.2.12
langchain_core==0.2.29
langchain-openai==0.1.20
openai==1.40.1
python-dotenv
streamlit==1.37.1